Log Insights: Refactor query sample handling to its own dedicated file
This is in preparation of a future commit that allows for more
sophisticated handling of the contents of query samples.
